REVISIONS TO color.tcl
1.  Remove option -variable
2.  Remove variable _varName which held the value of option -variable
3.  Use variable _unsavedSelection in place of $_varName when a value must be stored
4.  Add option -command
5.  Add variable _command which holds the value of option -command
6.  Set the default value for option -help to 1
7.  Add command _userCommand to evaluate $_command at stack level #0.
    _userCommand is called whenever the selected color changes, i.e. instead of
    setting $_varName.
8.  To avoid issues with trace, fully qualify ::SelectColor::_unsavedSelection
    when its value is set or it is used in a trace command.
9.  Use 8.4-compatible syntax for string indices in "string range" commands.
10. Amend SelectColor::_SetWithoutTrace so it sets the colors in the Color
    Selectors when the value in the entry widget changes.

PURPOSE
Tracing the variable set by the "-variable" option is too awkward in practice,
and so the "-variable" option has been replaced with "-command" which allows the
caller to specify a command to be executed whenever the selected color changes.

The use of the -command option is now included in the demo.

The properties of DynamicHelp balloon help have been changed in the demo, to
make the text easier to read.

The manual page has been updated to reflect the changes in color.tcl.